Add files via upload
cmon my man

+++ libft.h
@@ -18,27 +18,27 @@
 # include <unistd.h>
 # include <string.h>
 
-int	ft_isalpha(int c);
-int	ft_isdigit(int c);
-int	ft_isalnum(int c);
-int	ft_isascii(int c);
-int ft_isprint(int c);
+int		ft_isalpha(int c);
+int		ft_isdigit(int c);
+int		ft_isalnum(int c);
+int		ft_isascii(int c);
+int		ft_isprint(int c);
 size_t	ft_strlen(const char *str);
 void	*ft_memset(void *str, int c, size_t n);
 void	ft_bzero(void *str, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
-unsigned int	ft_strlcat(char *dest, const char *src, size_t size);
-int	ft_toupper(int c);
-int	ft_tolower(int c);
+size_t	ft_strlcat(char *dest, const char *src, size_t size);
+int		ft_toupper(int c);
+int		ft_tolower(int c);
 char	*ft_strchr(const char *str, int c);
 char	*ft_strrchr(char *str, int c);
-int	ft_strncmp(const char *s1, const char *s2, size_t n);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
 void	*memchr(const void *str, int c, size_t n);
-int	ft_memcmp(const void *str1, const void *str2, size_t n);
+int		ft_memcmp(const void *str1, const void *str2, size_t n);
 char	*ft_strnstr(const char *big, const char *little, size_t len);
-int	ft_atoi(char *str);
+int		ft_atoi(char *str);
 void	*calloc(size_t nitems, size_t size);
 char	*ft_strdup(const char *src);
